Create dept_first_search
The recursive version of the depth-first search algorithm in Python. We’ll use our TreeNode class to store the collection of values to search through. We’ll begin by implementing the basic recursive algorithm to search for a node with a specified value. Then we’ll extend the recursive function to also find the path from the root node to the target node.